libgolang: Provide custom bzero, mode_t and unistd.h on MSVC
We need unistd.h because e.g. `cimport posix.stat` forces inclusion of unistd.h
even if we use part of posix.stat that is available everywhere.

We use bzero in libgolang.cpp - provide it in similar way via compat
strings.h because on windows there is only string.h (no "s" in the end).

We also use mode_t in several places - provide it via typedef in
libgolang.h directly, because fcntl.h is present on MSVC.